refactor: restore viz_mpl and add a separate viz_plotly
